+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## CHANGELOG
 
+### v0.2.0
+
+#### Breaking change
+- the config keys being used, please refer the same config and rename the key attributes.
+
+#### Adds
+- ability to read container image to be read from config for a component being updated, for cases when deployment name
+is not the same as the container name for the container for which update is being made.
+
 ### v0.1.0
 
 OSS release of the internal tool which we have